Switched to double precision

+++ OpenGL1/Fragment_Shader.glsl
@@ -5,23 +5,23 @@ vec3 getColor(int iterations);
 layout(location = 0) out vec4 out_color;
 layout(location = 0) in vec2 position;
 
-uniform vec4 rect;					// vec4(-2.7, 1.0, 1.4, -1.4)
+uniform dvec4 rect;					// dvec4(-2.7, 1.0, 1.4, -1.4)
 uniform int maxiters;				// = 100;
 
 
-vec2 point;
+dvec2 point;
 
 void main(void)
 {
-	float dx = (rect.y - rect.x) / 2.0;					//(right - left) / 2.0  =  1.6
-	float dy = (rect.z - rect.w) / 2.0;					//(top - bottom) / 2.0  =  1.4
-	float midx = (rect.y + rect.x) / 2.0;				//(right + left) / 2.0  = -0.6
-	float midy = (rect.z + rect.w) / 2.0;				//(top + bottom) / 2.0  =  0
-	point = vec2((position[0] * dx) + midx, (position[1] * dy) + midy);
+	double dx = (rect.y - rect.x) / 2.0;					//(right - left) / 2.0  =  1.6
+	double dy = (rect.z - rect.w) / 2.0;					//(top - bottom) / 2.0  =  1.4
+	double midx = (rect.y + rect.x) / 2.0;				//(right + left) / 2.0  = -0.6
+	double midy = (rect.z + rect.w) / 2.0;				//(top + bottom) / 2.0  =  0
+	point = dvec2((position[0] * dx) + midx, (position[1] * dy) + midy);
 	
-	float temp;
+	double temp;
 	int iteration = 0;
-	vec2 val = point;
+	dvec2 val = point;
 	while(iteration<maxiters)             //actual iteration-count is i-1 at the end
 	{
 		if(val[0]*val[0] + val[1]*val[1] > 4)
